fix bug where navbar was disappearing when resizing the page
Replaced navbar element as previous one had too nav elements encapsulated within each other, which was causing the navbar links to disappear when page was resized to mimic a smaller screen. Furthermore, it looks like from what I can see here - https://getbootstrap.com/docs/4.0/components/navbar/ - and here - https://stackoverflow.com/questions/48320543/bootstrap-4-navbar-disappears-when-resizing-screen - the use of container-fluid class in the navbar would cause it to strech across the entire screen, which might also make it disappear or display in a strange manner. The navbar now in place switches from displaying the links horizontally to a hamburger menu on screen resize, inside which the links can be accessed.