use poetrydb api random poem endpoint to get a random poem
When the link "find me a poem" is clicked from within the navbar, a request to the poetrydb api random endpoint is launched, and a random poem is retrieved. The random poem title, author and lines fields are then being logged into the console. Next step is to create a render function so poem is displayed on the page.